Crate ipmpsc

Inter-Process Multiple Producer, Single Consumer Channels for Rust

This library provides a type-safe, high-performance inter-process channel implementation based on a shared memory ring buffer. It uses bincode for (de)serialization, including zero-copy deserialization, making it ideal for messages with large &str or &[u8] fields. And it has a name that rolls right off the tongue.

Modules

error

Structs

Receiver
Represents the receiving end of an inter-process channel, capable of receiving any message type implementing serde::Deserialize.

Sender
Represents the sending end of an inter-process channel.

SharedRingBuffer
Represents a file-backed shared memory ring buffer, suitable for constructing a Receiver or Sender.

ZeroCopyContext
Borrows a Receiver for the purpose of doing zero-copy deserialization of messages containing references.